+++ R/epidataCS_methods.R
@@ -7,8 +7,8 @@
 ### CONTINUOUS SPATIO-temporal infectious disease case data
 ###
 ### Copyright (C) 2009-2015,2017-2019 Sebastian Meyer
-### $Revision: 2505 $
-### $Date: 2019-11-29 15:39:58 +0100 (Fri, 29. Nov 2019) $
+### $Revision: 2958 $
+### $Date: 2023-03-10 08:43:46 +0100 (Fri, 10. Mar 2023) $
 ################################################################################
 
 
@@ -275,7 +275,7 @@ permute.epidataCS <- function (x, what = c("time", "space"), keep)
 
 print.epidataCS <- function (x, n = 6L, digits = getOption("digits"), ...)
 {
-    print.epidataCS_header(
+    print_epidataCS_header(
         timeRange = c(x$stgrid$start[1L], x$stgrid$stop[nrow(x$stgrid)]),
         bbox = bbox(x$W),
         nBlocks = length(unique(x$stgrid$BLOCK)),
@@ -304,7 +304,7 @@ print.epidataCS <- function (x, n = 6L, digits = getOption("digits"), ...)
     invisible(x)
 }
 
-print.epidataCS_header <- function (timeRange, bbox, nBlocks, nTiles,
+print_epidataCS_header <- function (timeRange, bbox, nBlocks, nTiles,
                                     digits = getOption("digits"))
 {
     bboxtxt <- paste(
@@ -354,7 +354,7 @@ summary.epidataCS <- function (object, ...)
 
 print.summary.epidataCS <- function (x, ...)
 {
-    print.epidataCS_header(timeRange = x$timeRange, bbox = x$bbox,
+    print_epidataCS_header(timeRange = x$timeRange, bbox = x$bbox,
                            nBlocks = x$nBlocks, nTiles = length(x$tileTable))
     cat("Overall number of events: ", x$nEvents, " (",
         if (x$nTypes==1) "single type" else paste(x$nTypes, "types"),