+++ R/epidataCS_plot.R
@@ -6,8 +6,8 @@
 ### plot-method for "epidataCS" objects
 ###
 ### Copyright (C) 2009-2015 Sebastian Meyer
-### $Revision: 2100 $
-### $Date: 2018-04-12 16:51:42 +0200 (Thu, 12. Apr 2018) $
+### $Revision: 2963 $
+### $Date: 2023-03-13 23:24:10 +0100 (Mon, 13. Mar 2023) $
 ################################################################################
 
 
@@ -254,9 +254,11 @@ epidataCSplot_space <- function (x, subset, by = type, tiles = x$W, pop = NULL,
     points.args_pointwise$cex <- points.args_pointwise$cex * cex.fun(pointCounts)
 
     ## plot
+    ## Note: "epidataCS" always works with projected coords,
+    ##       so we can set asp/aspect and thus avoid sp::is.projected()
     if (is.null(pop)) {
         ## classical plotting system
-        if (!add) plot(tiles, ...)
+        if (!add) plot(tiles, asp = 1, ...)
         do.call("points", c(alist(x=eventCoordsTypesCounts[,1:2,drop=FALSE]),
                             points.args_pointwise))
         ## optionally add legends
@@ -334,7 +336,7 @@ epidataCSplot_space <- function (x, subset, by = type, tiles = x$W, pop = NULL,
             list(inside = list(fun = lattice::draw.key(legend.counts), corner = corner.counts))
         }
         ## create the plot
-        spplot(obj = tiles, zcol = pop,
+        spplot(obj = tiles, zcol = pop, aspect = "iso",
                sp.layout = c(list(layout.points), sp.layout),
                legend = c(legend.types, legend.counts), ...)
     }