+++ R/hhh4_simulate.R
@@ -6,8 +6,8 @@
 ### Simulate from a HHH4 model
 ###
 ### Copyright (C) 2012 Michaela Paul, 2013-2016,2018,2021 Sebastian Meyer
-### $Revision: 2720 $
-### $Date: 2021-07-07 17:07:54 +0200 (Wed, 07. Jul 2021) $
+### $Revision: 2939 $
+### $Date: 2023-03-01 09:24:32 +0100 (Wed, 01. Mar 2023) $
 ################################################################################
 
 
@@ -64,13 +64,13 @@ simulate.hhh4 <- function (object, # result from a call to hhh4
     ## CAVE: for an ri()-model, building the terms affects the .Random.seed,
     ## so doing that twice would yield different simulations than pre-1.16.2
     if (is.null(object$terms))
-        object$terms <- terms.hhh4(object)
+        object$terms <- terms(object)
 
     ## get fitted exppreds nu_it, phi_it, lambda_it (incl. offsets, t in subset)
     exppreds <- get_exppreds_with_offsets(object, subset = subset, theta = theta)
 
     ## extract overdispersion parameters (simHHH4 assumes psi->0 means Poisson)
-    model <- terms.hhh4(object)
+    model <- terms(object)
     psi <- splitParams(theta,model)$overdisp
     if (length(psi) > 1) # "NegBinM" or shared overdispersion parameters
         psi <- psi[model$indexPsi]