+++ R/zzz.R
@@ -2,6 +2,12 @@
 ### Hook functions for package start-up
 #######################################
 
+gpcWarning <- function ()
+    .Deprecated(msg = paste(dQuote("gpc.poly"), "methods are deprecated",
+                            "in package", sQuote("surveillance")))
+
+## might no longer be needed in the future:
+## https://en.wikipedia.org/wiki/General_Polygon_Clipper
 gpclibCheck <- function (fatal = TRUE)
 {
     gpclibOK <- surveillance.options("gpclib")
@@ -21,25 +27,18 @@ gpclibCheck <- function (fatal = TRUE)
 
 .onAttach <- function (libname, pkgname)
 {
-    ## Startup message
     VERSION <- packageVersion(pkgname, lib.loc=libname)
-    packageStartupMessage("This is ", pkgname, " ", VERSION, ". ",
-                          "For overview type ",
-                          sQuote(paste0("help(", pkgname, ")")), ".")
+    packageStartupMessage("This is ", pkgname, " ", VERSION, "; ",
+                          "see ", sQuote(paste0("package?", pkgname)), " or\n",
+                          "https://surveillance.R-Forge.R-project.org/",
+                          " for an overview.")
 
-    ## decide if we should run all examples (some take a few seconds)
-    allExamples <- if (interactive()) {
-        TRUE
-    } else { # R CMD check
-        ## only do all examples if a specific environment variable is set
-        ## (to any value different from "")
-        nzchar(Sys.getenv("_R_SURVEILLANCE_ALL_EXAMPLES_"))
-        ## CAVE: testing for _R_CHECK_TIMINGS_ as in surveillance < 1.9-1
-        ## won't necessarily skip long examples for daily checks on CRAN (see
-        ## https://stat.ethz.ch/pipermail/r-devel/2012-September/064812.html
-        ## ). For instance, the daily Windows checks run without timings.
+    if (!interactive()) { # particularly for R CMD check
+        ## skip long examples and disallow gpclib, unless:
+        allExamples <- nzchar(Sys.getenv("_R_SURVEILLANCE_ALL_EXAMPLES_"))
+        ## not using surveillance.options() as this would load gpclib already
+        .Options$allExamples$value <- .Options$gpclib$value <- allExamples
     }
-    surveillance.options(allExamples = allExamples)
 }