+++ demo/cost.R
@@ -1,7 +1,6 @@
 ## need a writable figs/ directory to save selected figures
-OUTDIR <- tempdir()
-dir.create(file.path(OUTDIR, "figs"))
-message("saving selected figures in ", OUTDIR, "/figs")
+OUTDIR <- if (interactive()) tempdir() else "."
+stopifnot(dir.create(file.path(OUTDIR, "figs")))
 
 ###################################################
 ### chunk number 1:
@@ -14,7 +13,7 @@ set.seed(1234)
 opendevice <- function(horizontal=TRUE,width=7,height=4,...) {
   #Do it for postscript instead  -- who uses postscript these days??
   args <- list(...)
-  args$file <- file.path(OUTDIR, sub(".pdf",".eps",args$file))
+  args$file <- file.path(OUTDIR, sub("\\.pdf$",".eps",args$file))
   args$width <- width
   args$height <- height
   args$horizontal <- FALSE
@@ -22,14 +21,6 @@ opendevice <- function(horizontal=TRUE,width=7,height=4,...) {
 
   par(mar=c(4,4,2,2))
 }
-
-
-
-###################################################
-### chunk number 2: K1
-###################################################
-data("ha")
-plot(aggregate(ha),main="Hepatitis A in Berlin 2001-2006")
 
 
 ###################################################
@@ -47,7 +38,6 @@ dev.off()
 sps <- sim.pointSource(p = 0.99, r = 0.5, length = 400,
      A = 1, alpha = 1, beta = 0, phi = 0, frequency = 1,
      state = NULL, K = 1.7)
-plot(sps,xaxis.years=FALSE)
 
 
 ###################################################
@@ -59,28 +49,12 @@ dev.off()
 
 
 ###################################################
-### chunk number 6: HAB662 eval=FALSE
-###################################################
-## ha.b662 <- algo.bayes(aggregate(ha), control = list(range = 209:290, b = 2, w = 6, alpha = 0.01))
-## plot(ha.b662, firstweek=1, startyear = 2005)
-
-
-###################################################
 ### chunk number 7:
 ###################################################
 ha.b662 <- algo.bayes(aggregate(ha), control = list(range = 209:290, b = 2, w = 6, alpha = 0.01))
-plot(ha.b662, firstweek=1, startyear = 2005)
 opendevice(file="figs/hab662.pdf")
 plot(ha.b662, firstweek=1, startyear = 2005,legend.opts=list(x="topleft",horiz=TRUE))
 dev.off()
-
-
-###################################################
-### chunk number 8: FACDC eval=FALSE
-###################################################
-## cntrl <- list(range = 300:400, m = 1, w = 3, b = 5, alpha = 0.01)
-## sps.cdc <- algo.cdc(sps, control = cntrl)
-## sps.farrington <- algo.farrington(sps, control = cntrl)
 
 
 ###################################################
@@ -99,13 +73,6 @@ par(mfcol = c(1, 2),cex=0.8)
 plot(sps.cdc, legend = NULL, xaxis.years=FALSE)
 plot(sps.farrington, legend = NULL, xaxis.years=FALSE)
 dev.off()
-
-
-###################################################
-### chunk number 11: CUSUM eval=FALSE
-###################################################
-## kh <- find.kh(ARLa=500,ARLr=7)
-## ha.cusum <- algo.cusum(aggregate(ha),control=list(k=kh$k,h=kh$h,m="glm",trans="rossi",range=209:290))
 
 
 ###################################################
@@ -173,19 +140,6 @@ surv.one <- function(outbrk) {
 }
 
 
-
-###################################################
-### chunk number 15:  eval=FALSE
-###################################################
-## #Apply function to one
-## surv.one <- function(outbrk) {
-##   algo.compare(algo.call(outbrk,control=aparv.control))
-## }
-##
-## algo.summary(lapply(outbrks, surv.one))
-##
-
-
 ###################################################
 ### chunk number 16: ALGOSUMMARY
 ###################################################
@@ -216,40 +170,32 @@ print(res,digits=3)
 
 
 ###################################################
-### chunk number 19: HA eval=FALSE
-###################################################
-## shp <- system.file("shapes/berlin.shp",package="surveillance")
-## ha <- disProg2sts(ha, map=maptools::readShapePoly(shp,IDvar="SNAME"))
-## plot(ha,type=observed ~ 1 | unit)
-##
-
-
-###################################################
 ### chunk number 20:
 ###################################################
+
+## import shapefile as "SpatialPolygonsDataFrame"
+shp <- system.file("shapes/berlin.shp",package="surveillance")
+if (requireNamespace("maptools")) { # deprecated package
+    map <- maptools::readShapePoly(shp, IDvar = "SNAME")
+} else { # replacement code
+    map <- sf::as_Spatial(sf::st_read(shp, stringsAsFactors = TRUE, quiet = TRUE))
+    row.names(map) <- as.character(map$SNAME)
+}
+## convert to "sts" class
+ha.sts <- disProg2sts(ha, map = map)
+## or simply load the prepared object from the package: data("ha.sts")
+
 opendevice(file="figs/ha-1unit.pdf",width=7,height=7)
 par(mar=c(0,0,0,0))
-shp <- system.file("shapes/berlin.shp",package="surveillance")
-ha <- disProg2sts(ha, map=maptools::readShapePoly(shp,IDvar="SNAME"))
-plot(ha,type=observed ~ 1 | unit)
-
+plot(ha.sts,type=observed ~ 1 | unit)
 dev.off()
-
-
-###################################################
-### chunk number 21: HA:MAP eval=FALSE
-###################################################
-## ha4 <- aggregate(ha[,c("pank","mitt","frkr","scho","chwi","neuk")],nfreq=13)
-## ha4.cusum <- cusum(ha4,control=list(k=1.5,h=1.75,m="glm",trans="rossi",range=52:73))
-## #ha4.b332 <- bayes(ha4,control=list(range=52:73,b=2,w=3,alpha=0.01/6))
-## plot(ha4.cusum,type=observed ~ time | unit)
 
 
 ###################################################
 ### chunk number 22:
 ###################################################
 opendevice(file="figs/ha-timeunit.pdf",width=7,height=5)
-ha4 <- aggregate(ha[,c("pank","mitt","frkr","scho","chwi","neuk")],nfreq=13)
+ha4 <- aggregate(ha.sts[,c("pank","mitt","frkr","scho","chwi","neuk")],nfreq=13)
 ha4.cusum <- cusum(ha4,control=list(k=1.5,h=1.75,m="glm",trans="rossi",range=52:73))
 #ha4.b332 <- bayes(ha4,control=list(range=52:73,b=2,w=3,alpha=0.01/6))
 plot(ha4.cusum,type=observed ~ time | unit)