+++ tests/testthat/test-nbOrder.R
@@ -26,8 +26,8 @@ nbmat <- structure(
     .Dimnames = rep.int(list(c("a", "b", "c", "d", "e")), 2L)
     )
 
-test_that("nbOrder() returns the validated matrix", if (requireNamespace("spdep")) {
-    expect_identical(suppressMessages(nbOrder(adjmat, maxlag=Inf)),
+test_that("nbOrder() returns the validated matrix", {
+    expect_identical(nbOrder(adjmat),
                      nbmat)
 })